Add option to only load a node instead of the whole body
This adds the "rootSelector" option. When set, Snap will not replace the
whole document body with the new loaded body, but rather only the first
node that matches the selector (with `document.querySelector`) with the
matching node of the loaded page.

This makes it possible to easily persist some elements (i.e. chat
widgets that are added to the DOM by external libraries) across
navigation.